requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multitasking==0.0.9
numpy==1.19.2
pandas==1.1.2
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
uWSGI==2.0.19.1
wrapt==1.12.1
yfinance==0.1.54
